CORE-5114 WireTransactionKryoSerializer
