+++ libs/ledger/ledger-common-impl/build.gradle
@@ -16,15 +16,6 @@ dependencies {
 
     implementation project(":libs:crypto:crypto-core")
 
-    testCompileOnly 'org.osgi:osgi.core'
-    testRuntimeOnly "org.apache.felix:org.apache.felix.framework:$felixVersion"
-    testImplementation "org.mockito.kotlin:mockito-kotlin:$mockitoKotlinVersion"
-
-    testImplementation "com.esotericsoftware:kryo:$kryoVersion"
     testImplementation project(":libs:crypto:merkle-impl")
-    testImplementation project(":libs:crypto:cipher-suite-impl")
-    testImplementation project(":libs:serialization:serialization-kryo")
-    testImplementation project(":libs:sandbox")
-    testImplementation project(":libs:serialization:serialization-checkpoint-api")
 
 }
